Add a generic problem summarizer.
This will summarize problems for everything that's not a dependency or a
disk space. It does this by giving up and making rpm do it.

Also add an abstract format_details to ProblemSummary to make it more
clear that the class can't be used on its own.

(cherry picked from commit 01e0419f8e60f9553a079685354db331c1688b2d)